@@ -1,7 +1,9 @@
-var App = Vue.use(VueRouter).extend({});
-var router = new VueRouter();
-var routes = require('./routes');
+import routes from'./routes';
 
-router.map(routes).redirect({'*': '/index'}).afterEach(function(arg) {
-    document.title = arg.to.name;
-}).start(App, '#app');
+const router = new VueRouter({routes});
+
+Vue.use(VueRouter);
+
+const App = new Vue({
+    router
+}).$mount('#app');
